Use find_library to detect libmysqlclient
Fedora packages libmysqlclient.so in the mysql subdirectory of the
system library directory, which means that passing -lmysqlclient
to the linker will result in a linking error.  This commit adds
logic to search for libmysqlclient.so on the current library paths
as well as the system-wide path /usr/lib/mysql.

Signed-off-by: Rich Mattes <richmattes@gmail.com>